+++ test/integration/connectors/env_setup/sql/sqlite/source/sqlite-schema.sql
@@ -1,0 +1,5 @@
+CREATE TABLE cars (
+    car_id INTEGER PRIMARY KEY,
+    brand TEXT NOT NULL,
+    price INTEGER NOT NULL
+);