+++ CHANGELOG.md
@@ -1,9 +1,15 @@
-
-## 0.1.1-dev
+## 0.1.1-dev2
 
 ### Enhancements
 
 * **Migrate Slack Source Connector to V2**
+
+## 0.1.1-dev1
+
+### Enhancements
+
+* **Update KDB.AI vectorstore integration to 1.4**
+* **Add sqlite and postgres source connectors**
 
 ## 0.1.0